facerecog_tensorflow_gcp/gcf_face_recog/requirements.txt
oauth2client
google-api-python-client
opencv-python
requests
pillow
numpy
google-cloud-pubsub
google-cloud-bigquery
